# splitter: optimize performance