--- a/repo/splitter/splitter_buzhash32.go
+++ b/repo/splitter/splitter_buzhash32.go
@@ -66,7 +66,6 @@ func (rs *buzhash32Splitter) NextSplitPoint(b []byte) int {
 		}
 
 		fastPathBytes += fp
-		b = b[fp:]
 	}
 
 	// if we're over the max size, split