--- a/repo/splitter/splitter_rabinkarp64.go
+++ b/repo/splitter/splitter_rabinkarp64.go
@@ -66,7 +66,6 @@ func (rs *rabinKarp64Splitter) NextSplitPoint(b []byte) int {
 		}
 
 		fastPathBytes += fp
-		b = b[fp:]
 	}
 
 	// if we're over the max size, split