@jkowalski 2021-08-25
This looks amazing. I intended to investigate splitting performance fo a long time.